a lot of refactoring and reorganizing + formik validation
