+++ client/src/App.js
@@ -1,12 +1,12 @@
 import React, { useContext } from 'react';
 import { Switch, Route } from 'react-router-dom';
-import SignIn from './components/SignIn';
-import SignUp from './components/SignUp';
-import CreateRecord from './components/CreateRecord';
-import MainPage from './components/MainPage';
+import SignIn from './components/Pages/SignIn';
+import SignUp from './components/Pages/SignUp';
+import CreateRecord from './components/Pages/CreateRecord';
+import MainPage from './components/Pages/MainPage';
 
 import { AppContext } from './Context';
-import AdminPage from './components/AdminPage';
+import AdminPage from './components/Pages/AdminPage';
 
 export default function App() {
   const { user } = useContext(AppContext);
