+++ client/src/components/common/SingleRecord/Record.js
