+++ client/src/components/common/SingleRecord/RecordPannelHeadings.js
