+++ client/src/components/common/Validation.js
@@ -1,0 +1,25 @@
+import * as Yup from 'yup';
+
+export const signUpValidationSchema = Yup.object({
+  firstName: Yup.string()
+    .max(15, 'Must be 15 characters or less')
+    .required('Required'),
+  lastName: Yup.string()
+    .max(20, 'Must be 20 characters or less')
+    .required('Required'),
+  email: Yup.string()
+    .email('Invalid email address')
+    .required('Required'),
+  password: Yup.string()
+    .min(3, 'Must be 3 characters or more')
+    .max(15, 'Must be 15 characters or less')
+});
+
+export const logInValidationSchema = Yup.object({
+  email: Yup.string()
+    .email('Invalid email address')
+    .required('Required'),
+  password: Yup.string()
+    .min(3, 'Must be 3 characters or more')
+    .max(15, 'Must be 15 characters or less')
+});
