+++ client/src/components/common/inputElements/FormikFieldInput.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import { Field } from 'formik';
+import { TextField } from '@material-ui/core';
+
+export default function FormikFieldInput({
+  name,
+  label,
+  autoFocus = false,
+  touched,
+  errors,
+  margin = ''
+}) {
+  return (
+    <Field
+      error={touched[name] && errors[name] ? true : null}
+      helperText={touched[name] && errors[name] ? errors[name] : null}
+      as={TextField}
+      name={name}
+      variant='outlined'
+      type={name}
+      fullWidth
+      margin={margin}
+      label={label}
+      autoFocus={autoFocus}
+    />
+  );
+}
