+++ client/src/utils/api.js
@@ -1,7 +1,9 @@
 import axios from 'axios';
 
 export default {
+  //User
   getUser: () => axios.get('/api/user'),
+  createUser: user => axios.post('api/user', user),
 
   // Categories
   createCategory: category => axios.post('/api/category', category),
